Add the arena allocator
The arena allocator is based off of the Vmem allocator:

	http://www.google.com/search?q=bonwick+vmem

This will be the basis for all memory allocation.  Right now, it does
not have integrated qcaches (slabs).  That will require some work with
the slab allocator.  You can build a jumbo page allocator, using a
helper that xallocs with an alignment, which is pretty cool.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>