slab: Prepare for hashing to find bufctls
The slab allocator has a long-standing TODO: BUF.  Instead of using a
hash table to lookup a large object, we just used storage in the object
itself.  This was okay, other than possible fragmentation effects, but
it meant that the slab allocator touched every object it tracked.  We'll
eventually need an option to have "NO_TOUCH" slab allocators, where they
do not touch the objects they are tracking.  To do that, we'll need a
hash table.

This commit switches the bufctl struct from a TAILQ to an SLIST, which
isn't a big deal, but is a little better for a hash chain.  More
importantly, it statically allocates all of the boot-strapping caches.
Once the kmem_cache has a hashtable built in, it'll be a large slab
object, which means we'll need all three bootstrap caches to allocate
one.  This also fixes a FOREACH freeing bug.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>