slab: Use a hashtable when looking up bufctls
It only took 7 years or so to take care of that TODO.  Note that this is
probably slower than the old approach, which was instant.  We'll have to
think about adjusting the hashtable as it grows, if it becomes an issue.

Similar to the arena, the hashtable isn't using hashtable.h since that
code relies on kmalloc, and we'll have similar allocation issues
(probably need to directly alloc contig pages from the base_arena, since
we can't use anything that relies on us).

Aside from being necessary for NO_TOUCH slabs, this will save a lot of
memory when it comes to fragmentation when we use slabs for arenas.
Consider kpages_arena, which will have qcaches of pages.  That will be
pulling from slabs of 1 - 8 pages.  The one-page slab allocator will
need to have obj_size = PGSIZE and align = PGSIZE.  If we don't use the
hash and instead have the uintptr_t blob per object, we'd need two pages
per one-page-object, essentially wasting half of our memory.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>